Add sorting functionality to main component
